back-end-101/lesson-5/pyproject.toml
[tool.poetry]
name = "lesson-5"
version = "0.1.0"
description = ""
authors = ["deflillusion <defl.illusion@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
